+++ R/archive/fda-fig_01.R
@@ -43,17 +43,17 @@ make_fig_01 <- function(df,
   assert_choice(u_trtdur, c("days", "weeks", "months", "years"))
   assert_flag_variables(df, saffl_var)
 
-  df <- df %>%
-    filter(.data[[saffl_var]] == "Y") %>%
-    df_explicit_na() %>%
+  df <- df |>
+    filter(.data[[saffl_var]] == "Y") |>
+    df_explicit_na() |>
     mutate(
       TRTDUR = lubridate::interval(lubridate::ymd_hms(.data[[trtsdtm_var]]), lubridate::ymd_hms(.data[[trtedtm_var]])),
-      TRTDUR = TRTDUR %>% as.numeric(u_trtdur),
+      TRTDUR = TRTDUR |> as.numeric(u_trtdur),
       AVALU = u_trtdur
-    ) %>%
-    filter(!is.na(TRTDUR)) %>%
-    select(all_of(c(id_var, arm_var)), TRTDUR) %>%
-    distinct() %>%
+    ) |>
+    filter(!is.na(TRTDUR)) |>
+    select(all_of(c(id_var, arm_var)), TRTDUR) |>
+    distinct() |>
     arrange(desc(TRTDUR))
 
   df$PT_PCT <- seq_len(nrow(df)) / nrow(df) * 100