+++ R/archive/fda-table_11.R
@@ -28,8 +28,8 @@
 #' adae <- random.cdisc.data::cadae
 #'
 #' set.seed(1)
-#' adae <- adae %>%
-#'   rename(FMQ01SC = SMQ01SC) %>%
+#' adae <- adae |>
+#'   rename(FMQ01SC = SMQ01SC) |>
 #'   mutate(
 #'     AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
 #'     FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
@@ -76,9 +76,9 @@ make_table_11 <- function(df,
   )
 
   if (return_ard) {
-    return(list(table = tbl, ard = ard))
+    list(table = tbl, ard = ard)
   } else {
-    return(tbl) # nocov
+    tbl # nocov
   }
 }
 
@@ -99,13 +99,13 @@ preproc_df_table_11 <- function(df,
   assert_flag_variables(df, saffl_var)
   assert_subset(toupper(fmq_scope), c("NARROW", "BROAD"))
 
-  id_dcsae <- df %>%
-    filter(DCSREAS == "ADVERSE EVENT") %>%
-    select(all_of(c(id_var))) %>%
+  id_dcsae <- df |>
+    filter(DCSREAS == "ADVERSE EVENT") |>
+    select(all_of(c(id_var))) |>
     unlist()
 
-  df <- df %>%
-    filter(.data[[saffl_var]] == "Y", .data[[id_var]] %in% id_dcsae, .data[[fmqsc_var]] == fmq_scope) %>%
+  df <- df |>
+    filter(.data[[saffl_var]] == "Y", .data[[id_var]] %in% id_dcsae, .data[[fmqsc_var]] == fmq_scope) |>
     df_explicit_na(na_level = na_level)
   df[[fmqnam_var]] <- with_label(df[[fmqnam_var]], paste0("FMQ (", tools::toTitleCase(tolower(fmq_scope)), ")"))
 
@@ -121,8 +121,8 @@ preproc_df_table_11 <- function(df,
 #' adae <- random.cdisc.data::cadae
 #'
 #' set.seed(1)
-#' adae <- adae %>%
-#'   rename(FMQ01SC = SMQ01SC) %>%
+#' adae <- adae |>
+#'   rename(FMQ01SC = SMQ01SC) |>
 #'   mutate(
 #'     AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
 #'     FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
@@ -196,8 +196,8 @@ ard_table_11 <- function(df,
 #' adae <- random.cdisc.data::cadae
 #'
 #' set.seed(1)
-#' adae <- adae %>%
-#'   rename(FMQ01SC = SMQ01SC) %>%
+#' adae <- adae |>
+#'   rename(FMQ01SC = SMQ01SC) |>
 #'   mutate(
 #'     AESER = sample(c("Y", "N"), size = nrow(adae), replace = TRUE),
 #'     FMQ01NAM = sample(c("FMQ1", "FMQ2", "FMQ3"), size = nrow(adae), replace = TRUE)
@@ -241,57 +241,6 @@ make_table_11_gtsummary <- function(df,
     overall_row = TRUE,
     label = list(overall = "Patients with at least one AE leading to discontinuation")
   )
-
-  tbl
-}
-
-#' @export
-#' @rdname tbl_make_table_11
-make_table_11_rtables <- function(df,
-                                  alt_counts_df = NULL,
-                                  show_colcounts = TRUE,
-                                  id_var = "USUBJID",
-                                  arm_var = "ARM",
-                                  saffl_var = "SAFFL",
-                                  fmqsc_var = "FMQ01SC",
-                                  fmqnam_var = "FMQ01NAM",
-                                  fmq_scope = "NARROW",
-                                  lbl_overall = NULL,
-                                  risk_diff = NULL,
-                                  prune_0 = TRUE,
-                                  na_level = "<Missing>",
-                                  annotations = NULL) {
-  df <- preproc_df_table_11(df, id_var, arm_var, saffl_var, fmqsc_var, fmqnam_var, fmq_scope, na_level)
-  alt_counts_df <- alt_counts_df_preproc(alt_counts_df, id_var, arm_var, saffl_var)
-
-  lyt <- basic_table_annot(show_colcounts, annotations) %>%
-    split_cols_by_arm(arm_var, lbl_overall, risk_diff) %>%
-    analyze_num_patients(
-      vars = id_var,
-      riskdiff = !is.null(risk_diff),
-      .stats = c("unique"),
-      .labels = c(unique = "Patients with at least one AE leading to discontinuation")
-    ) %>%
-    split_rows_by(
-      "AEBODSYS",
-      label_pos = "topleft",
-      split_label = obj_label(df$AEBODSYS)
-    ) %>%
-    summarize_num_patients(
-      var = id_var,
-      riskdiff = !is.null(risk_diff),
-      .stats = "unique",
-      .labels = c(unique = NULL)
-    ) %>%
-    count_occurrences(
-      vars = fmqnam_var,
-      drop = FALSE,
-      riskdiff = !is.null(risk_diff)
-    ) %>%
-    append_varlabels(df, fmqnam_var, indent = 1L)
-
-  tbl <- build_table(lyt, df = df, alt_counts_df = alt_counts_df)
-  if (prune_0) tbl <- prune_table(tbl)
 
   tbl
 }