+++ R/archive/fda-table_21.R
@@ -19,13 +19,13 @@
 #' @examples
 #' library(dplyr)
 #'
-#' adsl <- random.cdisc.data::cadsl %>%
+#' adsl <- random.cdisc.data::cadsl |>
 #'   mutate(AGEGR1 = as.factor(case_when(
 #'     AGE >= 17 & AGE < 65 ~ ">=17 to <65",
 #'     AGE >= 65 ~ ">=65",
 #'     AGE >= 65 & AGE < 75 ~ ">=65 to <75",
 #'     AGE >= 75 ~ ">=75"
-#'   )) %>% formatters::with_label("Age Group, years")) %>%
+#'   )) |> formatters::with_label("Age Group, years")) |>
 #'   formatters::var_relabel(
 #'     AGE = "Age, years"
 #'   )
@@ -54,8 +54,8 @@ make_table_21 <- function(df,
   assert_subset(c(vars, id_var, arm_var, saffl_var), names(df))
   assert_flag_variables(df, c(saffl_var, "ASER"))
 
-  df <- df %>%
-    filter(.data[[saffl_var]] == "Y") %>%
+  df <- df |>
+    filter(.data[[saffl_var]] == "Y") |>
     df_explicit_na()
 
   # For percentages calculations in case of N_s, add the overall observations
@@ -77,17 +77,17 @@ make_table_21 <- function(df,
   lyt <- basic_table_annot(show_colcounts, annotations)
 
   lyt <- if (!is.null(lbl_overall) && denom != "N_s") {
-    lyt %>% split_cols_by_arm(arm_var, lbl_overall)
+    lyt |> split_cols_by_arm(arm_var, lbl_overall)
   } else {
-    lyt %>% split_cols_by_arm(arm_var)
+    lyt |> split_cols_by_arm(arm_var)
   }
 
-  lyt <- lyt %>%
+  lyt <- lyt |>
     count_patients_with_event(
       id_var,
       filters = c("AESER" = "Y"),
       .labels = c(count_fraction = "Any SAE")
-    ) %>%
+    ) |>
     analyze(
       vars = vars,
       var_labels = lbl_vars,
@@ -98,7 +98,7 @@ make_table_21 <- function(df,
         df_denom = if (!is.null(alt_counts_df)) alt_counts_df else df
       ),
       show_labels = "visible"
-    ) %>%
+    ) |>
     append_topleft("Characteristic")
 
   tbl <- build_table(lyt, df = df, alt_counts_df = alt_counts_df)
@@ -125,7 +125,7 @@ a_count_occurrences_ser_ae <- function(df,
                                        denom = c("N_s", "N_col", "n"),
                                        id_var = "USUBJID",
                                        arm_var = "ARM") {
-  df <- df %>% filter(ASER == "Y")
+  df <- df |> filter(ASER == "Y")
   occurrences <- df[[.var]]
   ids <- factor(df[[id_var]])
   has_occurrence_per_id <- table(occurrences, ids) > 0
@@ -140,10 +140,10 @@ a_count_occurrences_ser_ae <- function(df,
     N_s = lapply(
       lvls,
       function(x) {
-        df_denom %>%
-          filter(.data[[.var]] == x, .data[[arm_var]] == df[[arm_var]][1]) %>%
-          select(all_of(id_var)) %>%
-          distinct() %>%
+        df_denom |>
+          filter(.data[[.var]] == x, .data[[arm_var]] == df[[arm_var]][1]) |>
+          select(all_of(id_var)) |>
+          distinct() |>
           nrow()
       }
     ),