+++ R/cardinal.R
@@ -1,6 +1,10 @@
-#' cardinal
+#' cardinalfda
 #'
 #' Implementation of FDA Safety Tables and Figures
 #'
 #' @keywords internal
+#' @importFrom cards ard_continuous
+#' @importFrom cardx ard_continuous_ci
+#' @importFrom dplyr .data
+#' @importFrom gtsummary tbl_summary
 "_PACKAGE"